+++ src/index.ts
@@ -1,4 +1,4 @@
-import { RLogin } from './api'
+import { RLogin } from './RLogin'
 export { RLOGIN_AUTH_TOKEN_LOCAL_STORAGE_KEY } from './constants'
 
 export default RLogin
